+++ examples/autoscaling/main.tf
@@ -3,7 +3,7 @@ provider "aws" {
 }
 
 locals {
-  name   = "postgresql"
+  name   = "advanced"
   region = "eu-west-1"
   tags = {
     Owner       = "user"
@@ -14,10 +14,6 @@ locals {
 ################################################################################
 # Supporting Resources
 ################################################################################
-
-resource "random_password" "master" {
-  length = 10
-}
 
 module "vpc" {
   source  = "terraform-aws-modules/vpc/aws"
@@ -45,17 +41,24 @@ module "aurora" {
   engine                = "aurora-postgresql"
   engine_version        = "11.9"
   instance_type         = "db.r5.large"
-  instance_type_replica = "db.t3.medium"
+  instance_type_replica = "db.t3.large"
 
   vpc_id                = module.vpc.vpc_id
   db_subnet_group_name  = module.vpc.database_subnet_group_name
   create_security_group = true
   allowed_cidr_blocks   = module.vpc.private_subnets_cidr_blocks
 
-  replica_count                       = 2
-  iam_database_authentication_enabled = true
-  password                            = random_password.master.result
-  create_random_password              = false
+  replica_count         = 1
+  replica_scale_enabled = true
+  replica_scale_min     = 1
+  replica_scale_max     = 5
+
+  monitoring_interval           = 60
+  iam_role_name                 = "${local.name}-enhanced-monitoring"
+  iam_role_use_name_prefix      = true
+  iam_role_description          = "${local.name} RDS enhanced monitoring IAM role"
+  iam_role_path                 = "/autoscaling/"
+  iam_role_max_session_duration = 7200
 
   apply_immediately   = true
   skip_final_snapshot = true
@@ -68,15 +71,21 @@ module "aurora" {
 }
 
 resource "aws_db_parameter_group" "example" {
-  name        = "${local.name}-aurora-db-postgres11-parameter-group"
+  name_prefix = "${local.name}-aurora-db-postgres11-parameter-group"
   family      = "aurora-postgresql11"
   description = "${local.name}-aurora-db-postgres11-parameter-group"
   tags        = local.tags
 }
 
 resource "aws_rds_cluster_parameter_group" "example" {
-  name        = "${local.name}-aurora-postgres11-cluster-parameter-group"
+  name_prefix = "${local.name}-aurora-postgres11-cluster-parameter-group"
   family      = "aurora-postgresql11"
   description = "${local.name}-aurora-postgres11-cluster-parameter-group"
   tags        = local.tags
+}
+
+module "disabled_aurora" {
+  source = "../../"
+
+  create_cluster = false
 }