+++ examples/serverless/main.tf
@@ -1,53 +1,63 @@
 provider "aws" {
-  region = "us-east-1"
+  region = local.region
 }
 
-######################################
-# Data sources to get VPC and subnets
-######################################
-data "aws_vpc" "default" {
-  default = true
+locals {
+  name   = "serverless"
+  region = "eu-west-1"
+  tags = {
+    Owner       = "user"
+    Environment = "dev"
+  }
 }
 
-data "aws_subnet_ids" "all" {
-  vpc_id = data.aws_vpc.default.id
+################################################################################
+# Supporting Resources
+################################################################################
+
+module "vpc" {
+  source  = "terraform-aws-modules/vpc/aws"
+  version = "~> 2"
+
+  name = local.name
+  cidr = "10.99.0.0/18"
+
+  azs              = ["${local.region}a", "${local.region}b", "${local.region}c"]
+  public_subnets   = ["10.99.0.0/24", "10.99.1.0/24", "10.99.2.0/24"]
+  private_subnets  = ["10.99.3.0/24", "10.99.4.0/24", "10.99.5.0/24"]
+  database_subnets = ["10.99.7.0/24", "10.99.8.0/24", "10.99.9.0/24"]
+
+  tags = local.tags
 }
 
-#############
-# RDS Aurora
-#############
-module "aurora" {
+################################################################################
+# RDS Aurora Module - PostgreSQL
+################################################################################
+
+module "aurora_postgresql" {
   source = "../../"
-  name   = "aurora"
 
-  # PostgreSQL
-  engine = "aurora-postgresql"
+  name              = "${local.name}-postgresql"
+  engine            = "aurora-postgresql"
+  engine_mode       = "serverless"
+  storage_encrypted = true
 
-  # MySQL
-  # engine = "aurora"
+  vpc_id                = module.vpc.vpc_id
+  subnets               = module.vpc.database_subnets
+  create_security_group = true
+  allowed_cidr_blocks   = module.vpc.private_subnets_cidr_blocks
 
-  engine_mode           = "serverless"
-  engine_version        = null
   replica_scale_enabled = false
   replica_count         = 0
 
-  backtrack_window = 10 # ignored in serverless
-
-  subnets             = data.aws_subnet_ids.all.ids
-  vpc_id              = data.aws_vpc.default.id
   monitoring_interval = 60
-  skip_final_snapshot = true
-  instance_type       = "db.r4.large" # ignored for serverless
+
   apply_immediately   = true
-  storage_encrypted   = true
+  skip_final_snapshot = true
 
-  # PostgreSQL
-  db_parameter_group_name         = aws_db_parameter_group.aurora_db_postgresql10_parameter_group.id
-  db_cluster_parameter_group_name = aws_rds_cluster_parameter_group.aurora_cluster_postgresql10_parameter_group.id
-
-  # MySQL
-  #  db_parameter_group_name         = aws_db_parameter_group.aurora_db_aurora56_parameter_group.id
-  #  db_cluster_parameter_group_name = aws_rds_cluster_parameter_group.aurora_cluster_aurora56_parameter_group.id
+  db_parameter_group_name         = aws_db_parameter_group.example_postgresql.id
+  db_cluster_parameter_group_name = aws_rds_cluster_parameter_group.example_postgresql.id
+  # enabled_cloudwatch_logs_exports = # NOT SUPPORTED
 
   scaling_configuration = {
     auto_pause               = true
@@ -58,47 +68,68 @@ module "aurora" {
   }
 }
 
-# PostgreSQL
-resource "aws_db_parameter_group" "aurora_db_postgresql10_parameter_group" {
-  name        = "test-postgresql10-parameter-group"
+resource "aws_db_parameter_group" "example_postgresql" {
+  name        = "${local.name}-aurora-db-postgres-parameter-group"
   family      = "aurora-postgresql10"
-  description = "test-postgresql10-parameter-group"
+  description = "${local.name}-aurora-db-postgres-parameter-group"
+  tags        = local.tags
 }
 
-resource "aws_rds_cluster_parameter_group" "aurora_cluster_postgresql10_parameter_group" {
-  name        = "test-postgresql10-cluster-parameter-group"
+resource "aws_rds_cluster_parameter_group" "example_postgresql" {
+  name        = "${local.name}-aurora-postgres-cluster-parameter-group"
   family      = "aurora-postgresql10"
-  description = "test-postgresql10-cluster-parameter-group"
+  description = "${local.name}-aurora-postgres-cluster-parameter-group"
+  tags        = local.tags
 }
 
-# MySQL
-#resource "aws_db_parameter_group" "aurora_db_aurora56_parameter_group" {
-#  name        = "test-aurora56-parameter-group"
-#  family      = "aurora5.6"
-#  description = "test-aurora56-parameter-group"
-#}
-#
-#resource "aws_rds_cluster_parameter_group" "aurora_cluster_aurora56_parameter_group" {
-#  name        = "test-aurora56-cluster-parameter-group"
-#  family      = "aurora5.6"
-#  description = "test-aurora56-cluster-parameter-group"
-#}
+################################################################################
+# RDS Aurora Module - MySQL
+################################################################################
 
-############################
-# Example of security group
-############################
-resource "aws_security_group" "app_servers" {
-  name        = "app-servers"
-  description = "For application servers"
-  vpc_id      = data.aws_vpc.default.id
+module "aurora_mysql" {
+  source = "../../"
+
+  name              = "${local.name}-mysql"
+  engine            = "aurora-mysql"
+  engine_mode       = "serverless"
+  storage_encrypted = true
+
+  vpc_id                = module.vpc.vpc_id
+  subnets               = module.vpc.database_subnets
+  create_security_group = true
+  allowed_cidr_blocks   = module.vpc.private_subnets_cidr_blocks
+
+  replica_scale_enabled = false
+  replica_count         = 0
+
+  monitoring_interval = 60
+
+  apply_immediately   = true
+  skip_final_snapshot = true
+
+  db_parameter_group_name         = aws_db_parameter_group.example_mysql.id
+  db_cluster_parameter_group_name = aws_rds_cluster_parameter_group.example_mysql.id
+  # enabled_cloudwatch_logs_exports = # NOT SUPPORTED
+
+  scaling_configuration = {
+    auto_pause               = true
+    min_capacity             = 2
+    max_capacity             = 16
+    seconds_until_auto_pause = 300
+    timeout_action           = "ForceApplyCapacityChange"
+  }
 }
 
-resource "aws_security_group_rule" "allow_access" {
-  type                     = "ingress"
-  from_port                = module.aurora.this_rds_cluster_port
-  to_port                  = module.aurora.this_rds_cluster_port
-  protocol                 = "tcp"
-  source_security_group_id = aws_security_group.app_servers.id
-  security_group_id        = module.aurora.this_security_group_id
+resource "aws_db_parameter_group" "example_mysql" {
+  name        = "${local.name}-aurora-db-mysql-parameter-group"
+  family      = "aurora-mysql5.7"
+  description = "${local.name}-aurora-db-mysql-parameter-group"
+  tags        = local.tags
 }
 
+resource "aws_rds_cluster_parameter_group" "example_mysql" {
+  name        = "${local.name}-aurora-mysql-cluster-parameter-group"
+  family      = "aurora-mysql5.7"
+  description = "${local.name}-aurora-mysql-cluster-parameter-group"
+  tags        = local.tags
+}