+++ examples/serverless/versions.tf
@@ -1,7 +1,10 @@
 terraform {
-  required_version = ">= 0.12.6"
+  required_version = ">= 0.13"
 
   required_providers {
-    aws = ">= 3.8"
+    aws = {
+      source  = "hashicorp/aws"
+      version = ">= 3.8"
+    }
   }
 }