+++ versions.tf
@@ -1,8 +1,15 @@
 terraform {
-  required_version = ">= 0.12.6"
+  required_version = ">= 0.13"
 
   required_providers {
-    aws    = ">= 3.8"
-    random = ">= 2.2"
+    aws = {
+      source  = "hashicorp/aws"
+      version = ">= 3.30"
+    }
+
+    random = {
+      source  = "hashicorp/random"
+      version = ">= 2.2"
+    }
   }
 }